Minor changes to matrix classes

+++ Application.cpp
@@ -147,25 +147,25 @@ void Application::initialize() {
 	start = clock();
 	{
 		{
-			PolynomialMatrix<double> pm = PolynomialMatrix<double>(3, 3);
+			PolynomialMatrix<float> pm = PolynomialMatrix<float>(3, 3);
 
-			//double numArr[] = { 5, 3, 2, -2, 1, 4, 2, 5, -2 };
-			double numArr[] = { 5, 3, 9, -2, 1, 3, 2, 5, 15 };
-			std::vector<double> values2(numArr,
+			//float numArr[] = { 5, 3, 2, -2, 1, 4, 2, 5, -2 };
+			float numArr[] = { 5, 3, 9, -2, 1, 3, 2, 5, 15 };
+			std::vector<float> values2(numArr,
 					numArr + sizeof(numArr) / sizeof(numArr[0]));
 			pm.setVector(values2);
 
-			double denArr[] = { 5, 1 };
-			std::vector<double> numCoeffs(denArr,
+			float denArr[] = { 5, 1 };
+			std::vector<float> numCoeffs(denArr,
 					denArr + sizeof(denArr) / sizeof(denArr[0]));
 
-			Polynomial<double> p = Polynomial<double>(numCoeffs, 1);
+			Polynomial<float> p = Polynomial<float>(numCoeffs, 1);
 			pm.set(0, 0, p);
 
 			pm.print();
 			pm.upperTriangular().print();
 			pm.rref().print();
-			std::vector<std::vector<Polynomial<double> > > specialSolutions = pm.specialSolutions();
+			std::vector<std::vector<Polynomial<float> > > specialSolutions = pm.specialSolutions();
 			for (int i = 0; i < specialSolutions.size(); i++) {
 				std::cout << "Special solution " << i + 1 << ": ";
 				for (int j = 0; j < specialSolutions[i].size(); j++)
@@ -179,18 +179,18 @@ void Application::initialize() {
 				<< std::endl;
 
 		{
-			ComplexMatrix<double> cm = ComplexMatrix<double>(3, 3);
+			ComplexMatrix<float> cm = ComplexMatrix<float>(3, 3);
 
-			double numArr[] = { 5, 3, 2, -2, 1, 4, 2, 5, -2 };
-			std::vector<double> values2(numArr,
+			float numArr[] = { 5, 3, 2, -2, 1, 4, 2, 5, -2 };
+			std::vector<float> values2(numArr,
 					numArr + sizeof(numArr) / sizeof(numArr[0]));
 			cm.setVector(values2);
 
-			double denArr[] = { 5, 1 };
-			std::vector<double> numCoeffs(denArr,
+			float denArr[] = { 5, 1 };
+			std::vector<float> numCoeffs(denArr,
 					denArr + sizeof(denArr) / sizeof(denArr[0]));
 
-			Complex<double> c = Complex<double>(numCoeffs, 1);
+			Complex<float> c = Complex<float>(numCoeffs, 1);
 			cm.set(0, 0, c);
 
 			cm.print();
@@ -201,23 +201,23 @@ void Application::initialize() {
 		std::cout << std::endl << "-------------------" << std::endl
 				<< std::endl;
 
-		Matrix<double> m1 = Matrix<double>(3, 3);
-		//double arr[] = { 4, -4, 8, -4, 4, -8, 8, -8, 16 };
-		double arr[] = { 5, 3, 2, -2, 1, 4, 2, 5, -2 };
-		//double arr[] = { 5, 25, 2, -2, -10, 4, 2, 10, -2 };
-		//double arr[] = { 5, 0, 2, -2, 0, 4, 2, 0, -2 };
-		std::vector<double> values(arr, arr + sizeof(arr) / sizeof(arr[0]));
+		Matrix<float> m1 = Matrix<float>(3, 3);
+		//float arr[] = { 4, -4, 8, -4, 4, -8, 8, -8, 16 };
+		float arr[] = { 5, 3, 2, -2, 1, 4, 2, 5, -2 };
+		//float arr[] = { 5, 25, 2, -2, -10, 4, 2, 10, -2 };
+		//float arr[] = { 5, 0, 2, -2, 0, 4, 2, 0, -2 };
+		std::vector<float> values(arr, arr + sizeof(arr) / sizeof(arr[0]));
 		m1.setVector(values);
 
 		m1.print();
-		std::vector<Complex<double> > eigenvalues;
+		std::vector<Complex<float> > eigenvalues;
 		eigenvalues = m1.eigenvalues();
 		for (int i = 0; i < eigenvalues.size(); i++) {
 			std::cout << "Eigenvalue " << i + 1 << ": ";
 			eigenvalues[i].print();
 		}
 
-		std::vector<std::vector<Complex<double> > > evectors = m1.eigenvectors();
+		std::vector<std::vector<Complex<float> > > evectors = m1.eigenvectors();
 		for (int i = 0; i < evectors.size(); i++) {
 			std::cout << "Eigenvector " << i + 1 << ": ";
 			for (int j = 0; j < evectors[i].size(); j++)
@@ -228,10 +228,10 @@ void Application::initialize() {
 		std::cout << std::endl << "-------------------" << std::endl
 				<< std::endl;
 
-		m1 = Matrix<double>(3, 3);
-		double arr2[] = { 5, 25, 2, -2, -10, 4, 2, 10, -2 };
-		//double arr2[] = { 5, 0, 2, -2, 0, 4, 2, 0, -2 };
-		std::vector<double> values2(arr2, arr2 + sizeof(arr2) / sizeof(arr2[0]));
+		m1 = Matrix<float>(3, 3);
+		float arr2[] = { 5, 25, 2, -2, -10, 4, 2, 10, -2 };
+		//float arr2[] = { 5, 0, 2, -2, 0, 4, 2, 0, -2 };
+		std::vector<float> values2(arr2, arr2 + sizeof(arr2) / sizeof(arr2[0]));
 		m1.setVector(values2);
 
 		m1.print();
@@ -252,8 +252,8 @@ void Application::initialize() {
 		std::cout << std::endl << "-------------------" << std::endl
 				<< std::endl;
 
-		double arr3[] = { 5, 25, 10, -2, -10, -4, 2, 10, 4 };
-		std::vector<double> values3(arr3, arr3 + sizeof(arr3) / sizeof(arr3[0]));
+		float arr3[] = { 5, 25, 10, -2, -10, -4, 2, 10, 4 };
+		std::vector<float> values3(arr3, arr3 + sizeof(arr3) / sizeof(arr3[0]));
 		m1.setVector(values3);
 
 		m1.print();
@@ -274,7 +274,7 @@ void Application::initialize() {
 		std::cout << std::endl << "-------------------" << std::endl
 				<< std::endl;
 
-		m1 = Matrix<double>::identity(3);
+		m1 = Matrix<float>::identity(3);
 
 		m1.print();
 		eigenvalues = m1.eigenvalues();
